Convert FooterBlock into RSC
