+++ apps/store/src/services/storyblok/commonStoryblokComponents.ts
@@ -12,7 +12,7 @@ import { ContactSupportBlock } from '@/blocks/ContactSupportBlock'
 import { ContentBlock } from '@/blocks/ContentBlock'
 import { CookieListBlock } from '@/blocks/CookieListBlock'
 import { DownloadableContentItemBlock } from '@/blocks/DownloadableContentItemBlock'
-import { FooterBlock, FooterLinkBlock, FooterSectionBlock } from '@/blocks/FooterBlock'
+import { FooterBlock, FooterLinkBlock, FooterSectionBlock } from '@/blocks/FooterBlock/FooterBlock'
 import { GridBlock } from '@/blocks/GridBlock'
 import { HeaderBlock } from '@/blocks/HeaderBlock/HeaderBlock'
 import { NavItemBlock } from '@/blocks/HeaderBlock/NavItemBlock'
